requirement.txt
# Core Django and DRF
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3

# Filtering and pagination
django-filter==24.2

# Database
psycopg2-binary==2.9.9  # For PostgreSQL support

# Cloud storage
google-cloud-storage==2.10.0

# Development tools
django-debug-toolbar==4.2.0
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.0

# Monitoring and logging
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
